Egistec: Use safer synchronized worker thread.
The old implementation has relied on the Android service that
is supposed to wait on callbacks before initiating another event.
However, logs indicate that it is possible to receive a cancel() call
followed by authenticate() before onError(canceled) is reported back to
the system. This messes up the state stored in the eventfd by implicit
addition on write, for which reason the synchronized worker keeps it in
a simple variable protected by a mutex.

This simplification also means dropping one slightly similar worker
thread implementation, allowing for a single one that is tested and
maintained on all platforms.
Muxing eventfd and fingerprint still happens in a separate class, to be
able to share this (now minimal) implementation between Nile and Ganges
HALs.
